migrate 'latest_at' test over to re_query
